Try out organization structure for banner tests
This is just a proof of concept that deletes Tests 04-11 from
mobile-action-button-leads-to-fullpage-banner. Please do not merge this!

Add testFeature method to Banner class, to bundle test steps and give
thhem names.

The logging does not work properly - the feature test "marker" log
entries from `testFeature` should be interspersed with the log entries
form the test steps